+++ cli/src/commands/agent/run/mode_interactive.rs
@@ -204,16 +204,9 @@ pub async fn run_interactive(
             None
         };
 
-        // Convert CLI CommandsConfig to TUI CommandsConfig
-        let commands_config_for_tui =
-            config
-                .commands_config
-                .as_ref()
-                .map(|c| stakpak_tui::CommandsConfig {
-                    include: c.include.clone(),
-                    exclude: c.exclude.clone(),
-                    definitions: c.definitions.clone(),
-                });
+        // CommandsConfig is now the same type in CLI and TUI (from stakpak_shared)
+        let commands_config_for_tui = config.commands_config.clone();
+
         let tui_handle = tokio::spawn(async move {
             let latest_version = get_latest_cli_version().await;
             stakpak_tui::run_tui(
