+++ libs/api/Cargo.toml
@@ -26,6 +26,7 @@ tokio = { workspace = true, features = ["macros"] }
 serde_yaml = "0.9.34"
 once_cell = "1.21.3"
 regex.workspace = true
+dirs = "5.0"
 
 [dev-dependencies]
 tempfile = { workspace = true }
