+++ libs/api/src/slash_commands/explain.md
