+++ libs/api/src/slash_commands/predefined.rs
