+++ libs/shared/Cargo.toml
@@ -34,6 +34,7 @@ rustls-platform-verifier = { workspace = true }
 russh = { version = "0.53.0" }
 russh-sftp = { version = "2.1.1" }
 dirs = "5.0"
+glob = { workspace = true }
 async-trait = { workspace = true }
 futures = { workspace = true }
 futures-util = { workspace = true }
